Fix artifacts in the boundary box of Volumetrics
Strangely I found sometimes Alpha value goes over 1.0!!  That's not normal and probably the cause of the artifacts. For now, I've added a check in integrator.cc to make sure alpha never goes over 1.0, but we should investigate the root cause of why alpha goes over 1.0 in the first place.

 Changes to be committed:
	modified:   src/yafraycore/integrator.cc